+++ src/endpoints/mod.rs
@@ -1,3 +1,4 @@
 pub(crate) mod settings;
 pub(crate) mod net;
 pub(crate) mod debug;
+pub(crate) mod i18n;
